# workspaces/x2a/plugins/x2a-backend/src/schema/openapi.yaml
openapi: 3.0.4
info:
  title: x2a
  description: The [X2Ansible](https://github.com/redhat-developer/x2ansible) API.
  version: 0.0.1
servers:
  - url: /
paths:
...
  /rules:
    get:
      summary: Returns a list of all rules.
      responses:
        '200':
          description: All rules.
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Rule'
    post:
      summary: Creates a new rule (admin only).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  description: Title of the rule
                description:
                  type: string
                  description: Description of the rule
                required:
                  type: boolean
                  description: Whether the rule is required for all projects
              required:
                - title
                - description
      responses:
        '201':
          description: Created rule.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rule'
  /rules/{ruleId}:
    get:
      summary: Returns a rule by ID.
      parameters:
        - in: path
          name: ruleId
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Rule data.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rule'
        '404':
          description: Rule not found.
    put:
      summary: Updates a rule by ID (admin only).
      parameters:
        - in: path
          name: ruleId
          schema:
            type: string
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  description: Title of the rule
                description:
                  type: string
                  description: Description of the rule
                required:
                  type: boolean
                  description: Whether the rule is required for all projects
              required:
                - title
                - description
                - required
      responses:
        '200':
          description: Updated rule.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rule'
        '404':
          description: Rule not found.
    delete:
      summary: Deletes a rule by ID (admin only).
      parameters:
        - in: path
          name: ruleId
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Rule deleted successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  deletedCount:
                    type: integer
                    description: Number of rules deleted (0 or 1)
                required:
                  - deletedCount
        '404':
          description: Rule not found.
...
components:
  securitySchemes:
    callbackSignature:
      type: apiKey
      in: header
      name: X-Callback-Signature
      description: |
        HMAC-SHA256 signature of raw request body using job-specific callback token.
        Signature is computed as: HMAC-SHA256(callbackToken, raw_request_body)
        and sent as a hex-encoded string.

        Replay attack prevention: Requests are only accepted within 3 hours of job creation time.
  schemas:
...
    Rule:
      type: object
      properties:
        id:
          type: string
          description: UUID for the rule
        title:
          type: string
          description: Title of the rule
        description:
          type: string
          description: Description of the rule
        required:
          type: boolean
          description: Whether the rule is required for all projects
        createdAt:
          type: string
          format: date-time
          description: Date/time when the rule was created
        updatedAt:
          type: string
          format: date-time
          description: Date/time when the rule was last updated
      required:
        - id
        - title
        - description
        - required
        - createdAt
        - updatedAt
...